RIFEX-51 Pay to RNS domain
* Fixing callbacks and actions
