+++ app/scripts/controllers/rif/constants/lumino.js
@@ -1,0 +1,24 @@
+import {CALLBACKS} from '@rsksmart/lumino-light-client-sdk/dist/utils/callbacks';
+
+export const lumino = {
+  callbacks: {
+    RECEIVED_PAYMENT: CALLBACKS.RECEIVED_PAYMENT,
+    COMPLETED_PAYMENT: CALLBACKS.COMPLETED_PAYMENT,
+    EXPIRED_PAYMENT: CALLBACKS.EXPIRED_PAYMENT,
+    OPEN_CHANNEL: CALLBACKS.OPEN_CHANNEL,
+    REQUEST_CLIENT_ONBOARDING: CALLBACKS.REQUEST_CLIENT_ONBOARDING,
+    CLIENT_ONBOARDING_SUCCESS: CALLBACKS.CLIENT_ONBOARDING_SUCCESS,
+    DEPOSIT_CHANNEL: CALLBACKS.DEPOSIT_CHANNEL,
+    CLOSE_CHANNEL: CALLBACKS.CLOSE_CHANNEL,
+    REQUEST_OPEN_CHANNEL: CALLBACKS.REQUEST_OPEN_CHANNEL,
+    REQUEST_DEPOSIT_CHANNEL: CALLBACKS.REQUEST_DEPOSIT_CHANNEL,
+    REQUEST_CLOSE_CHANNEL: CALLBACKS.REQUEST_CLOSE_CHANNEL,
+    SENT_PAYMENT: CALLBACKS.SENT_PAYMENT,
+    SIGNING_FAIL: CALLBACKS.SIGNING_FAIL,
+    FAILED_OPEN_CHANNEL: CALLBACKS.FAILED_OPEN_CHANNEL,
+    FAILED_DEPOSIT_CHANNEL: CALLBACKS.FAILED_DEPOSIT_CHANNEL,
+    FAILED_CLOSE_CHANNEL: CALLBACKS.FAILED_CLOSE_CHANNEL,
+    FAILED_PAYMENT: CALLBACKS.FAILED_PAYMENT,
+    FAILED_CREATE_PAYMENT: CALLBACKS.FAILED_CREATE_PAYMENT,
+  },
+}
