+++ app/scripts/controllers/rif/lumino/callbacks.js
@@ -1,18 +1,17 @@
 import {CALLBACKS} from '@rsksmart/lumino-light-client-sdk/dist/utils/callbacks';
 
 export class LuminoCallbacks {
+
+  callbackNames = Object.keys(CALLBACKS).map(callbackKey => CALLBACKS[callbackKey]);
+
   constructor (lumino) {
     this.lumino = lumino;
   }
 
-  getAvailableCallbacks () {
-    return Promise.resolve(Object.keys(CALLBACKS));
-  }
-
   listenForCallback (callbackName) {
     return new Promise((resolve, reject) => {
-      if (CALLBACKS[callbackName]) {
-        this.lumino.callbacks.set(CALLBACKS[callbackName], (result, error) => {
+      if (this.callbackNames.indexOf(name => callbackName === name)) {
+        this.lumino.callbacks.set(callbackName, (result, error) => {
           if (error) {
             return reject(error);
           }
