+++ ui/app/rif/actions/index.js
@@ -1,8 +1,8 @@
 import * as niftyActions from '../../actions';
 import extend from 'xtend';
-import {buildActions} from './lumino';
+import {lumino} from '../../../../app/scripts/controllers/rif/constants';
 
-let rifActions = {
+const rifActions = {
   SHOW_MODAL: 'SHOW_MODAL',
   SHOW_MENU: 'SHOW_MENU',
   NAVIGATE_TO: 'NAVIGATE_TO',
@@ -35,6 +35,15 @@ let rifActions = {
   getDomains,
   getDomain,
   updateDomains,
+  // Lumino
+  onboarding,
+  openChannel,
+  closeChannel,
+  getChannels,
+  getAvailableCallbacks,
+  listenCallback,
+  createPayment,
+  createDeposit,
 }
 
 let background = null;
@@ -495,6 +504,124 @@ function updateDomains (domain) {
   };
 }
 
-rifActions = buildActions(background, niftyActions, rifActions);
+// Lumino
+
+function onboarding () {
+  return (dispatch) => {
+    return new Promise((resolve, reject) => {
+      background.rif.lumino.onboarding((error) => {
+        if (error) {
+          dispatch(niftyActions.displayWarning(error));
+          return reject(error);
+        }
+        dispatch(this.listenCallback(lumino.callbacks.CLIENT_ONBOARDING_SUCCESS))
+          .then(result => {
+            dispatch(niftyActions.displayToast('Client successfully onboarded!'));
+          }).catch(error => {
+          dispatch(niftyActions.displayWarning(error));
+        });
+        return resolve();
+      });
+    });
+  };
+}
+
+function listenCallback (callbackName) {
+  return (dispatch) => {
+    return new Promise((resolve, reject) => {
+      background.rif.lumino.listenCallback(callbackName, (result, error) => {
+        if (error) {
+          dispatch(niftyActions.displayWarning(error));
+          return reject(error);
+        }
+        return resolve(result);
+      });
+    });
+  };
+}
+
+function getAvailableCallbacks () {
+  return (dispatch) => {
+    return new Promise((resolve, reject) => {
+      background.rif.lumino.getAvailableCallbacks((callbackNames, error) => {
+        if (error) {
+          dispatch(niftyActions.displayWarning(error));
+          return reject(error);
+        }
+        return resolve(callbackNames);
+      });
+    });
+  };
+}
+
+function openChannel (partner, tokenAddress) {
+  return (dispatch) => {
+    return new Promise((resolve, reject) => {
+      background.rif.lumino.openChannel(partner, tokenAddress, (error) => {
+        if (error) {
+          dispatch(niftyActions.displayWarning(error));
+          return reject(error);
+        }
+        return resolve();
+      });
+    });
+  };
+}
+
+function closeChannel (partner, tokenAddress, address, tokenNetworkAddress, channelIdentifier) {
+  return (dispatch) => {
+    return new Promise((resolve, reject) => {
+      background.rif.lumino.closeChannel(partner, tokenAddress, address, tokenNetworkAddress, channelIdentifier, (error) => {
+        if (error) {
+          dispatch(niftyActions.displayWarning(error));
+          return reject(error);
+        }
+        return resolve();
+      });
+    });
+  };
+}
+
+function createDeposit (partner, tokenAddress, address, tokenNetworkAddress, channelIdentifier, netAmount) {
+  return (dispatch) => {
+    return new Promise((resolve, reject) => {
+      background.rif.lumino.createDeposit(partner, tokenAddress, address, tokenNetworkAddress, channelIdentifier, netAmount, (error) => {
+        if (error) {
+          dispatch(niftyActions.displayWarning(error));
+          return reject(error);
+        }
+        return resolve();
+      });
+    });
+  };
+}
+
+function createPayment (partner, tokenAddress, netAmount) {
+  return (dispatch) => {
+    return new Promise((resolve, reject) => {
+      background.rif.lumino.createPayment(partner, tokenAddress, netAmount, (error) => {
+        if (error) {
+          dispatch(niftyActions.displayWarning(error));
+          return reject(error);
+        }
+        return resolve();
+      });
+    });
+  };
+}
+
+function getChannels () {
+  return (dispatch) => {
+    return new Promise((resolve, reject) => {
+      background.rif.lumino.getChannels((channels, error) => {
+        if (error) {
+          dispatch(niftyActions.displayWarning(error));
+          return reject(error);
+        }
+        return resolve(channels);
+      });
+    });
+  };
+}
 
 module.exports = rifActions
